compiler warnings removed

+++ code/qcommon/vm_local.h
@@ -131,7 +131,7 @@ struct vm_s {
     // DO NOT MOVE OR CHANGE THESE WITHOUT CHANGING THE VM_OFFSET_* DEFINES
     // USED BY THE ASM CODE
     int			programStack;		// the vm may be recursively entered
-    int			(*systemCall)( int *parms );
+    intptr_t	(*systemCall)( intptr_t *parms );
 
 	//------------------------------------
